Extract shared contentSurfaceShape into common components
