+++ src/com/android/messaging/ui/conversationpicker/common/PickerTokens.kt
@@ -1,19 +1,5 @@
 package com.android.messaging.ui.conversationpicker.common
 
-import androidx.compose.foundation.shape.CornerBasedShape
-import androidx.compose.foundation.shape.CornerSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.ReadOnlyComposable
 import androidx.compose.ui.unit.dp
 
 internal val ScreenContentPadding = 8.dp
-
-private val ZeroCornerSize = CornerSize(0.dp)
-
-internal val MaterialTheme.contentSurfaceShape: CornerBasedShape
-    @Composable @ReadOnlyComposable
-    get() = shapes.large.copy(
-        bottomStart = ZeroCornerSize,
-        bottomEnd = ZeroCornerSize,
-    )
